Add loading to history grid & up next
